Update README.md
Didn't do a update in a while, i've been busy with college stuff and so it's been hard to work on the mod. Now, i will focus on the new character sprite sheet (You won't have to battle against a placeholder anymore), as well as the music. Anyway, this repo will be made public when i merge this branch (pls, no stealing or i will make this private again). See ya.